Add tests to prepare for refactoring PCA/RDA/KDE model
- export offline_analysis and return figure handles from
  generate_offline_analysis_screen() to allow testing
- Add integration tests in bcipy/signal/tests/model/test_offline_analysis.py,
  asserting model's current behavior for:
  - AUC (number only, no contents of pkl file)
  - lik_dens figure
  - mean_erp figure
- Add unit tests in bcipy/signal/tests/model/test_model.py,
  asserting model's current behavior
- Remove unused grid_search() function from cross_validation.py
- Add placeholder tests for notch and bandpass
- Add some notes in model training code
- Removing unnecessary folders, formatting, and removing unused code from demos
- Remove unnecessary DummyDimReduction
- Rename "mach_learning" to "ml

TODO - it might be worth considering deleting or updating the demos at
bcipy/signal/model/demo/ml
