+++ bcipy/helpers/demo/demo_eeg_model_related.py
@@ -1,6 +1,6 @@
 import numpy as np
 from bcipy.helpers.copy_phrase_wrapper import CopyPhraseWrapper
-from bcipy.signal.model.mach_learning.train_model import train_pca_rda_kde_model
+from bcipy.signal.model.ml.train_model import train_pca_rda_kde_model
 from bcipy.helpers.task import alphabet
 
 channel_map = [0] + [1] * 16 + [0, 0, 1, 1, 0, 1, 1, 1, 0]
