+++ bcipy/signal/model/__init__.py
@@ -1,5 +1,7 @@
+from .offline_analysis import offline_analysis
 from .model import SignalModel
 
 __all__ = [
     "SignalModel",
+    "offline_analysis",
 ]
