+++ dev_requirements.txt
@@ -6,3 +6,4 @@ mockito==1.1.0
 pylint==2.3.1
 pytest==3.1.3
 twine==3.2.0
+pytest-mpl
